made game and board class, removed extra files, removed extra print and console.log from files
