+++ mysite/checkers/consumers.py
@@ -3,9 +3,8 @@ import json
 from asgiref.sync import async_to_sync
 from channels.generic.websocket import WebsocketConsumer
 import logging
-from .checkers_game_engine import *
 logger = logging.getLogger("mylogger")
-from .rules import Game
+from .game import Game
 
 #this class is about websocket communication
 # when websocket is connected disconnects and message is received respective fuctions is triggered 
@@ -21,7 +20,7 @@ class GameConsumer(WebsocketConsumer):
         )
 		#call some function here which return board according to the room and player who is requesting
         logger.info('connected to websocket')
-        # board = play_display_checkers(UserPlayer, UserPlayer, upper_color=State.WHITE)
+		
         global game
         game = Game()
         game.event_loop()
@@ -33,7 +32,8 @@ class GameConsumer(WebsocketConsumer):
                 'type': 'game_message',
                 'message': board,
 				'moves': moves,
-				'selected_piece' : selected_piece
+				'selected_piece' : selected_piece,
+				'turn': game.turn,
             }
         )
         self.accept()
@@ -51,16 +51,16 @@ class GameConsumer(WebsocketConsumer):
         text_data_json = json.loads(text_data)
         message = text_data_json['message']
         logger.info(text_data)
-        #do some thing here
+        #click is recieved here are update board is sent back
         global game
         game.event_loop(message)
         board, moves, selected_piece = game.update()
-        print (board)
         self.send(text_data=json.dumps({
 		                     'message': board, 
 		                     'moves': moves, 
 		                     'selected_piece' : selected_piece,
-		                     'room_name':self.room_name}))
+		                     'room_name':self.room_name,
+							 'turn': game.turn,}))
         
         # Send message to room group
         # async_to_sync(self.channel_layer.group_send)(
@@ -74,7 +74,6 @@ class GameConsumer(WebsocketConsumer):
     # Receive message from room group
     def game_message(self, event):
         message = event['message']
-        print (event)
         # Send message to WebSocket
         self.send(text_data=json.dumps(event))
         logger.info('game_message funcion')
